TD-3895-DataService reference removed and required service classes created in the web project.
